# doc/swagger.yml
openapi: 3.0.1
info:
  title: dating server API
  description: Описание методов API сервера мобильного приложения
  version: 0.24.6
servers:
  - url: https://dev.dating.eagledev.ru/api/v1
  - url: https://prod.dating.eagledev.ru/api/v1
  - url: http://localhost:8032/api/v1/
paths:
...
  /like:
    get:
      tags:
        - Лайки
      summary: Список лайков
      security:
        - bearerAuth: [ ]
      description: "список пользователей, которые поставили лайк. Для экрана Вы по\
        нравились"
      responses:
        "200":
          description: "OK, массив профилей"
          content:
            application/json:
              schema:
                required:
                  - data
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: успех запроса
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Couple'
        "500":
          description: server error
          content: { }
...
  /find:
    post:
      tags:
        - Поиск пары
      summary: Поиск пар
      security:
        - bearerAuth: [ ]
      description: запрос на получение пар с учетом фильтра и лимита кол-ва
      requestBody:
        content:
          application/json:
            schema:
              required:
                - filter
                - limit
              type: object
              properties:
                filter:
                  type: object
                  properties:
                    sex:
                      type: string
                      description: enum пол (male\female\any)
                      example: male
                    radius:
                      type: integer
                      description: радиус поиска по геолокации (в километрах). 0 -
                        отключение фильтра.
                      example: 100
                    ageFrom:
                      type: integer
                      description: возраст От
                      example: 18
                    ageTo:
                      type: integer
                      description: возраст До
                      example: 25
                  description: фильтры для поиска пар
                limit:
                  type: integer
                  description: кол-во пар в ответе
                  example: 10
        required: true
      responses:
        "200":
          description: "OK, массив Пар"
          content:
            application/json:
              schema:
                required:
                  - data
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: успех запроса
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Couple'
        "403":
          description: 1017 geo location not set
          content:
            application/json:
              schema:
                required:
                  - error
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: успех запроса
                    example: false
                  error:
                    required:
                      - code
                      - message
                    type: object
                    properties:
                      code:
                        type: integer
                        description: code
                        example: 1017
                      message:
                        type: string
                        description: message
                        example: user geolocation is not set
        "500":
          description: server error
          content: { }
      x-codegen-request-body-name: data
  /find/like:
    post:
      tags:
        - Поиск пары
      summary: Поставить лайк
      security:
        - bearerAuth: [ ]
      description: отправка лайка (по uuid Пользователя)
      requestBody:
        content:
          application/json:
            schema:
              required:
                - userID
              type: object
              properties:
                userID:
                  type: string
                  description: uuid Пользователя (которому поставить лайк)
                  example: 6a7485ca-2020-470d-b0b5-04f3ff19f12f
        required: true
      responses:
        "200":
          description: "OK, лайк поставлен"
          content:
            application/json:
              schema:
                required:
                  - data
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: успех запроса
                    example: true
                  data:
                    required:
                      - match
                    type: object
                    properties:
                      match:
                        type: boolean
                        description: совпадение лайков
                        example: false
                      matchID:
                        type: string
                        description: "опциональное поле. если math true, вернётся\
                          \ uuid мэтча"
                        example: uuid
        "500":
          description: server error
          content: { }
      x-codegen-request-body-name: data
  /find/dislike:
    post:
      tags:
        - Поиск пары
      summary: Поставить дизлайк
      security:
        - bearerAuth: [ ]
      description: отправка дизайлка (по uuid Пользователя)
      requestBody:
        content:
          application/json:
            schema:
              required:
                - userID
              type: object
              properties:
                userID:
                  type: string
                  description: uuid Пользователя (которому поставить дизлайк)
                  example: 6a7485ca-2020-470d-b0b5-04f3ff19f12f
        required: true
      responses:
        "200":
          description: "OK, дизлайк поставлен"
          content:
            application/json:
              schema:
                required:
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: успех запроса
                    example: true
        "500":
          description: server error
          content: { }
      x-codegen-request-body-name: data
...
  /find/clear-likes:
    get:
      tags:
        - Debug
      summary: Удалить свои лайки\дизлайки
      security:
        - bearerAuth: [ ]
      description: "Очистит лайки и дизлайки, которые ставил пользователь. нужно д\
        ля отладки и тестов. Работает только для авторизованнх пользователей"
      responses:
        "200":
          description: "OK, удалено"
          content:
            application/json:
              schema:
                required:
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: успех запроса
                    example: true
        "500":
          description: server error
          content: { }
...
components:
  schemas:
...
    UserProfile:
      required:
        - ID
      type: object
      properties:
        ID:
          type: string
          description: uuid пользователя
        name:
          type: string
          description: имя
        sex:
          type: string
          description: пол
        birthday:
          type: string
          description: дата рождения
        city:
          type: string
          description: город
        about:
          type: string
          description: о себе
        photos:
          type: array
          items:
            $ref: '#/components/schemas/UserPhoto'
        distance:
          type: integer
          description: расстояние до пользователя в километрах
          example: 4389
    UserPhoto:
      required:
        - ID
        - URL
        - isMain
      type: object
      properties:
        ID:
          type: string
          description: uuid фотографии
          example: 6a7485ca-2020-470d-b0b5-04f3ff19f12f
        URL:
          type: string
          description: URL адрес картинки
          example: https://example.example/example.jpg
        isMain:
          type: boolean
          description: флаг основное фото профиля
          example: false
        createdAt:
          type: string
          description: дата добавления
    Couple:
      required:
        - distance
        - profile
      type: object
      properties:
        distance:
          type: integer
          description: расстояние до пары (в километрах)
          example: 5
        profile:
          $ref: '#/components/schemas/UserProfile'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT